fix admin registration using abx hooks
