+++ archivebox/api/apps.py
@@ -2,10 +2,14 @@ __package__ = 'archivebox.api'
 
 from django.apps import AppConfig
 
+import abx
 
 
 class APIConfig(AppConfig):
     name = 'api'
 
-    def ready(self):
-        pass
+
+@abx.hookimpl
+def register_admin(admin_site):
+    from api.admin import register_admin
+    register_admin(admin_site)
