+++ archivebox/core/models.py
@@ -43,9 +43,11 @@ from ..extractors import ARCHIVE_METHODS_INDEXING_PRECEDENCE, EXTRACTORS
 
 
 
+
+
 class Tag(ABIDModel):
     """
-    Based on django-taggit model + ABID base.
+    Loosely based on django-taggit model + ABID base.
     """
     abid_prefix = 'tag_'
     abid_ts_src = 'self.created_at'
